Changed playbackPeriodDefault and updated conductor with a music player
Changed playbackPeriodDefault to give the audio shield more time to play. I think it was getting out of sync too quickly with a shorter callback time.
Conductor.ino now has a music player, still only gets data from the randomCart() function, but it plays it well when all the correct files are on the SD card